+++ src/rex.i18n/src/rex/i18n/middleware.py
@@ -5,8 +5,7 @@
 
 from rex.web import Pipe
 
-from .core import get_i18n_context, KEY_LOCALE, KEY_TIMEZONE, \
-    get_locale_identifier
+from .core import get_i18n_context
 from .extensions import LocaleDetector, TimezoneDetector
 
 
@@ -25,17 +24,12 @@ class I18NPipe(Pipe):
     def establish_locale(self, request):
         i18n = get_i18n_context()
         if not i18n.has_locale():
-            i18n.set_locale(LocaleDetector.detect_locale(request))
-        locale = i18n.get_locale()
-        request.environ['rex.session'][KEY_LOCALE] = \
-            get_locale_identifier(locale)
+            i18n.set_locale(LocaleDetector.detect_locale(request), request)
 
     def establish_timezone(self, request):
         i18n = get_i18n_context()
         if not i18n.has_timezone():
             i18n.set_timezone(TimezoneDetector.detect_timezone(request))
-        timezone = i18n.get_timezone()
-        request.environ['rex.session'][KEY_TIMEZONE] = timezone.zone
 
     def cleanup(self):
         i18n = get_i18n_context()